vision_service/requirements.txt
fastapi
uvicorn
pillow
numpy
torch
torchvision
timm
python-multipart